bot/requirements.txt
python-telegram-bot==20.3
aiohttp==3.8.4
psutil==5.9.5
